collapse functions from easybuild.base.missing into easybuild.tools.utilities
