+++ easybuild/base/frozendict.py
@@ -1,4 +1,4 @@
-### External compatible license
+# External compatible license
 # taken from https://github.com/slezica/python-frozendict on March 14th 2014 (commit ID b27053e4d1)
 #
 # Copyright (c) 2012 Santiago Lezica
@@ -22,6 +22,8 @@ It can be used as a drop-in replacement for dictionaries where immutability is d
 """
 import operator
 from collections import Mapping
+
+from easybuild.base import fancylogger
 
 
 # minor adjustments:
@@ -56,3 +58,44 @@ class FrozenDict(Mapping):
     # minor adjustment: define missing keys() method
     def keys(self):
         return self.__dict.keys()
+
+
+class FrozenDictKnownKeys(FrozenDict):
+    """A frozen dictionary only allowing known keys."""
+
+    # list of known keys
+    KNOWN_KEYS = []
+
+    def __init__(self, *args, **kwargs):
+        """Constructor, only way to define the contents."""
+        self.log = fancylogger.getLogger(self.__class__.__name__, fname=False)
+
+        # support ignoring of unknown keys
+        ignore_unknown_keys = kwargs.pop('ignore_unknown_keys', False)
+
+        # handle unknown keys: either ignore them or raise an exception
+        tmpdict = dict(*args, **kwargs)
+        unknown_keys = [key for key in tmpdict.keys() if key not in self.KNOWN_KEYS]
+        if unknown_keys:
+            if ignore_unknown_keys:
+                for key in unknown_keys:
+                    self.log.debug("Ignoring unknown key '%s' (value '%s')" % (key, args[0][key]))
+                    # filter key out of dictionary before creating instance
+                    del tmpdict[key]
+            else:
+                msg = "Encountered unknown keys %s (known keys: %s)" % (unknown_keys, self.KNOWN_KEYS)
+                self.log.raiseException(msg, exception=KeyError)
+
+        super(FrozenDictKnownKeys, self).__init__(tmpdict)
+
+    # pylint: disable=arguments-differ
+    def __getitem__(self, key, *args, **kwargs):
+        """Redefine __getitem__ to provide a better KeyError message."""
+        try:
+            return super(FrozenDictKnownKeys, self).__getitem__(key, *args, **kwargs)
+        except KeyError as err:
+            if key in self.KNOWN_KEYS:
+                raise KeyError(err)
+            else:
+                tup = (key, self.__class__.__name__, self.KNOWN_KEYS)
+                raise KeyError("Unknown key '%s' for %s instance (known keys: %s)" % tup)
