+++ easybuild/base/generaloption.py
@@ -44,10 +44,9 @@ from optparse import SUPPRESS_HELP as nohelp  # supported in optparse of python 
 from optparse import gettext as _gettext  # this is gettext.gettext normally
 
 from easybuild.base.fancylogger import getLogger, setroot, setLogLevel, getDetailsLogLevels
-from easybuild.base.missing import nub
 from easybuild.base.optcomplete import autocomplete, CompleterOption
 from easybuild.tools.docs import mk_rst_table
-from easybuild.tools.utilities import shell_quote
+from easybuild.tools.utilities import nub, shell_quote
 
 
 HELP_OUTPUT_FORMATS = ['', 'rst', 'short', 'config']
