+++ easybuild/tools/config.py
@@ -43,7 +43,7 @@ import tempfile
 import time
 
 from easybuild.base import fancylogger
-from easybuild.base.missing import FrozenDictKnownKeys
+from easybuild.base.frozendict import FrozenDictKnownKeys
 from easybuild.base.patterns import Singleton
 from easybuild.tools.build_log import EasyBuildError
 from easybuild.tools.module_naming_scheme import GENERAL_CLASS
