+++ easybuild/tools/filetools.py
@@ -55,11 +55,11 @@ import zlib
 from xml.etree import ElementTree
 
 from easybuild.base import fancylogger
-from easybuild.base.missing import nub
+from easybuild.tools import run
 # import build_log must stay, to use of EasyBuildLog
 from easybuild.tools.build_log import EasyBuildError, dry_run_msg, print_msg
 from easybuild.tools.config import build_option
-from easybuild.tools import run
+from easybuild.tools.utilities import nub
 
 try:
     import requests
