+++ easybuild/tools/job/backend.py
@@ -32,9 +32,8 @@ Abstract interface for submitting jobs and related utilities.
 from abc import ABCMeta, abstractmethod
 
 from easybuild.base import fancylogger
-from easybuild.base.missing import get_subclasses
 from easybuild.tools.config import get_job_backend
-from easybuild.tools.utilities import import_available_modules
+from easybuild.tools.utilities import get_subclasses, import_available_modules
 
 
 class JobBackend(object):
