+++ easybuild/tools/package/utilities.py
@@ -37,7 +37,6 @@ import tempfile
 import pprint
 
 from easybuild.base import fancylogger
-from easybuild.base.missing import get_subclasses
 from easybuild.base.patterns import Singleton
 from easybuild.tools.config import PKG_TOOL_FPM, PKG_TYPE_RPM, build_option, get_package_naming_scheme, log_path
 from easybuild.tools.build_log import EasyBuildError
@@ -45,7 +44,9 @@ from easybuild.tools.filetools import change_dir, which
 from easybuild.tools.package.package_naming_scheme.pns import PackageNamingScheme
 from easybuild.tools.run import run_cmd
 from easybuild.tools.toolchain import DUMMY_TOOLCHAIN_NAME
-from easybuild.tools.utilities import import_available_modules
+from easybuild.tools.utilities import get_subclasses, import_available_modules
+
+
 _log = fancylogger.getLogger('tools.package')  # pylint: disable=C0103
