+++ easybuild/tools/repository/repository.py
@@ -35,9 +35,8 @@ Generic support for dealing with repositories
 :author: Fotis Georgatos (Uni.Lu, NTUA)
 """
 from easybuild.base import fancylogger
-from easybuild.base.missing import get_subclasses
 from easybuild.tools.build_log import EasyBuildError
-from easybuild.tools.utilities import import_available_modules
+from easybuild.tools.utilities import get_subclasses, import_available_modules
 
 _log = fancylogger.getLogger('repository', fname=False)
 
@@ -147,7 +146,7 @@ def avail_repositories(check_useable=True):
 
     class_dict = dict([(x.__name__, x) for x in get_subclasses(Repository) if x.USABLE or not check_useable])
 
-    if not 'FileRepository' in class_dict:
+    if 'FileRepository' not in class_dict:
         raise EasyBuildError("avail_repositories: FileRepository missing from list of repositories")
 
     return class_dict
